feat: add ability to trace handshake
In order to do that, we refactor the configuration handling. We move
options to openvpn options, and we create a config object that is
broader in scope.

This commit also updates the binary entrypoint to test this
functionality.

Since I was here, I added the pinger too to exercise the pinger over the
tunnel.

+++ internal/reliabletransport/packets.go
@@ -19,7 +19,7 @@ type inFlightPacket struct {
 	packet *model.Packet
 
 	// retries is a monotonically increasing counter for retransmission.
-	retries uint8
+	retries int
 }
 
 func newInFlightPacket(p *model.Packet) *inFlightPacket {
